+++ Courier/src/utopia/courier/controller/read/EmailReader.scala
@@ -2,11 +2,13 @@ package utopia.courier.controller.read
 
 import utopia.courier.model.Email
 import utopia.courier.model.read.DeletionRule.NeverDelete
-import utopia.courier.model.read.{DeletionRule, ReadSettings}
+import utopia.courier.model.read.{DeletionRule, FolderPath, ReadSettings}
+import utopia.flow.collection.CollectionExtensions._
+import utopia.flow.collection.immutable.caching.LazyTree
 import utopia.flow.operator.EqualsExtensions._
 import utopia.flow.parse.AutoClose._
 import utopia.flow.parse.AutoCloseWrapper
-import utopia.flow.collection.CollectionExtensions._
+import utopia.flow.view.immutable.caching.Lazy
 import utopia.flow.view.mutable.caching.ResettableLazy
 
 import java.io.InputStream
@@ -14,8 +16,8 @@ import java.nio.file.Path
 import java.util.Properties
 import javax.mail.Message.RecipientType
 import javax.mail.internet.MimeMessage
-import javax.mail.{Flags, Folder, Message, Multipart, Part, Session}
-import scala.collection.immutable.VectorBuilder
+import javax.mail._
+import scala.annotation.tailrec
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
 
@@ -103,60 +105,60 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 {
 	/**
 	  * Asynchronously reads and parses email
-	  * @param folderName Name of the folder to read (default = INBOX)
+	  * @param targetFolders Targeted folder or folders. Default = Inbox.
 	  * @param maxMessageCount Maximum number of messages processed at this time. Negative if not limited (default).
 	  * @param skipMessageCount Number of messages to skip from the beginning (default = 0)
 	  * @param deletionRule A rule to apply to deleting messages (default = never delete messages)
 	  * @param exc Implicit execution context
 	  * @return A future with the read results, when they come available
 	  */
-	def apply(folderName: String = "INBOX", maxMessageCount: Int = -1, skipMessageCount: Int = 0,
-	          deletionRule: DeletionRule = NeverDelete)(implicit exc: ExecutionContext) =
-		Future { readBlocking(folderName, maxMessageCount, skipMessageCount, deletionRule) }
+	def apply(targetFolders: TargetFolders = TargetFolders.inbox, maxMessageCount: Int = -1, skipMessageCount: Int = 0,
+	          deletionRule: DeletionRule = NeverDelete)
+	         (implicit exc: ExecutionContext) =
+		Future { readBlocking(targetFolders, maxMessageCount, skipMessageCount, deletionRule) }
 	
 	/**
 	  * Asynchronously iterates over read email
-	  * @param folderName Name of the folder to read (default = INBOX)
-	  * @param maxMessageCount Maximum number of messages processed at this time. Negative if not limited (default).
+	  * @param targetFolders Targeted folder or folders. Default = Inbox.
 	  * @param skipMessageCount Number of messages to skip from the beginning (default = 0)
 	  * @param deletionRule A rule to apply to deleting messages (default = never delete messages)
-	  * @param f Function that accepts the message iterator. The iterator will terminate after a failure.
+	  * @param f Function that accepts an iterator that returns read messages.
+	 *                   The specified iterator may contain any number of failures.
 	  * @param exc Implicit execution context
 	  * @return A future with the read results, when they come available
 	  */
-	def iterateAsync[B](folderName: String = "INBOX", maxMessageCount: Int = -1, skipMessageCount: Int = 0,
+	def iterateAsync[B](targetFolders: TargetFolders = TargetFolders.inbox, skipMessageCount: Int = 0,
 	                    deletionRule: DeletionRule = NeverDelete)(f: Iterator[Try[A]] => B)
 	                   (implicit exc: ExecutionContext) =
-		Future { iterateBlocking(folderName, maxMessageCount, skipMessageCount, deletionRule)(f) }
+		Future { iterateBlocking(targetFolders, skipMessageCount, deletionRule)(f) }
 	
 	/**
-	  * Reads email data from the targeted message folder. Processes read data.
-	  * @param folderName Name of the folder to read (default = INBOX)
+	  * Reads email data from the targeted message folder.
+	  * @param targetFolders Targeted folder or folders. Default = Inbox.
 	  * @param maxMessageCount Maximum number of read messages (includes those skipped).
 	  *                        Negative if not limited (default)
 	  * @param skipMessageCount Number of messages to skip from the beginning (default = 0)
 	  * @param deletionRule A rule to apply to deleting messages (default = never delete messages)
 	  * @return Newly read data. Failure if connection or read setup failed
 	  */
-	def readBlocking(folderName: String = "INBOX", maxMessageCount: Int = -1, skipMessageCount: Int = 0,
+	def readBlocking(targetFolders: TargetFolders = TargetFolders.inbox,
+	                 maxMessageCount: Int = -1, skipMessageCount: Int = 0,
 	                 deletionRule: DeletionRule = NeverDelete) =
-		iterateBlocking(folderName, maxMessageCount, skipMessageCount, deletionRule) {
-			_.tryMap { a => a }.map { _.toVector }
-		}
+		iterateBlocking(targetFolders, skipMessageCount, deletionRule) { _.take(maxMessageCount).toTry }
 	
 	/**
 	  * Reads and processes email data using a message iterator. Message reading is terminated if
 	  * reading or parsing fails.
-	  * @param folderName Name of the folder to read (default = INBOX)
-	  * @param maxMessageCount Maximum number of read messages (includes those skipped).
-	  *                        Negative if not limited (default)
+	  * @param targetFolders Targeted folder or folders. Default = Inbox.
 	  * @param skipMessageCount Number of messages to skip from the beginning (default = 0)
 	  * @param deletionRule A rule to apply to deleting messages (default = never delete messages)
-	  * @param f Function that accepts the message iterator. The iterator will terminate after a failure.
+	  * @param f Function that accepts an iterator that returns read messages.
+	 *          The specified iterator may contain any number of failures.
 	  * @return Newly read data. Failure if connection or read setup failed
 	  */
-	def iterateBlocking[B](folderName: String = "INBOX", maxMessageCount: Int = -1, skipMessageCount: Int = 0,
-	                       deletionRule: DeletionRule = NeverDelete)(f: Iterator[Try[A]] => B) =
+	def iterateBlocking[B](targetFolders: TargetFolders = TargetFolders.inbox,
+	                       skipMessageCount: Int = 0,
+	                       deletionRule: DeletionRule = NeverDelete)(f: Iterator[Try[A]] => B): B =
 	{
 		// Initializes the session
 		settings.modify(new Properties).flatMap { properties =>
@@ -168,83 +170,194 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 				.flatMap { _.tryConsumeContent { store =>
 					// Connects to the mail server
 					store.connect(settings.hostAddress, settings.authentication.user, settings.authentication.password)
-					// Reads the targeted mail folder
-					AutoCloseWrapper(store.getFolder(folderName)) { _.close(deletionRule.canDelete) }
-						.consumeContent { folder =>
-							Try {
-								// Opens the folder
-								folder.open(if (deletionRule.canDelete) Folder.READ_WRITE else Folder.READ_ONLY)
-								// Reads targeted messages in the folder
-								// (may skip some / limit amount of processed messages)
-								val folderSize = folder.getMessageCount
-								val maxReadIndex = if (maxMessageCount < 0) folderSize else
-									(skipMessageCount + maxMessageCount) min folderSize
-								val sourceIterator = {
-									if (skipMessageCount >= folderSize)
-										Iterator.empty
-									else
-										(skipMessageCount + 1 to maxReadIndex).iterator
-											.map { index => Try { folder.getMessage(index) } }
-								}
-								// Forms the message processing iterator and gives it to the specified function
-								val iterator = new MessageIterator(new RawMessageIterator(sourceIterator, makeBuilder))
-								val finalResult = f(iterator)
-								// Handles message deletion (ignores failures)
-								Try {
-									val (processed, skipped, failed) = iterator.popMessages()
-									failed.foreach { message => if (deletionRule.shouldDeleteFailed) deleteMessage(message) }
-									if (skipped.nonEmpty && deletionRule.shouldDeleteSkipped)
-										deleteMessages(skipped)
-									if (processed.nonEmpty && deletionRule.shouldDeleteProcessed(failed.isEmpty))
-										deleteMessages(processed)
-								}
-								finalResult
-							}
+					// Determines the targeted folders
+					val folderTree = LazyTree
+						.iterate(Lazy { store.getDefaultFolder -> FolderPath(Vector()) }) { case (folder, path) =>
+							folder.list().iterator.map { subFolder => Lazy { subFolder -> (path/subFolder.getName) } }
 						}
-				}.flatten }
+					// Prepares the folders for reading
+					val rawFoldersIterator = targetFolders(folderTree.map { _._2 })
+						// Resolves the resulting paths
+						.map { path =>
+							// Case: Root/default folder was targeted
+							if (path.isEmpty)
+								Success(store.getDefaultFolder)
+							// Case: A non-root folder was targeted => Determines the targeted folder
+							else
+								path.parts
+									// Traverses the tree using the specified path
+									.foldLeftIterator(Success(folderTree): Try[LazyTree[(Folder, FolderPath)]]) { (parent, nextPart) =>
+										parent.flatMap { p =>
+											p.children.find { _.nav._2.name ~== nextPart }
+												.toTry { new NoSuchElementException(
+													s"$p doesn't contain a folder named $nextPart") }
+										}
+									}
+									// Fails if targeting a non-existing folder
+									.takeTo { _.isFailure }
+									.last.map { _.nav._1 }
+						}
+					// Calls the specified function using a prepared message iterator
+					// Forms the message processing iterator and gives it to the specified function
+					val foldersIterator = new FoldersIterator(rawFoldersIterator, skipMessageCount, deletionRule)
+					val iterator = new MessageIterator(
+						new RawMessageIterator(foldersIterator, deletionRule, makeBuilder), deletionRule)
+					val result = f(iterator)
+					// Closes any folder that may have been left open
+					foldersIterator.close()
+					result
+				} }
 		// Handles failures by giving them to the specified function
 		}.getOrMap { error => f(Iterator.single(Failure(error))) }
 	}
 	
-	private def deleteMessage(message: Message) = message.setFlag(Flags.Flag.DELETED, true)
-	private def deleteMessages(messages: Iterable[Message]) = messages.foreach(deleteMessage)
+	private def delete(message: Message) = Try { message.setFlag(Flags.Flag.DELETED, true) }
 	
 	
 	// NESTED   -------------------------------------
 	
-	private class MessageIterator(source: RawMessageIterator) extends Iterator[Try[A]]
+	private class FoldersIterator(source: Iterator[Try[Folder]], skipMessageCount: Int, deletionRule: DeletionRule)
+		extends Iterator[Try[Message]]
 	{
 		// ATTRIBUTES   ---------------------------
 		
-		private val nextMaterials = ResettableLazy {
-			source.nextWhere {
-				case Success(msg) => msg.isDefined
-				case Failure(_) => true
-			}.map { _.map { _.get } }
-		}
-		private val processedMessagesBuilder = new VectorBuilder[Message]()
+		// Number of message skips yet to be performed
+		private var remainingSkipCount = skipMessageCount
+		// Contains currently open folder and its size
+		private var openFolder: Option[(Folder, Int)] = None
+		// Index of the next message to read - 1-based
+		private var nextMessageIndex = 1
+		private var queuedFailures: Iterator[Throwable] = Iterator.empty
 		
-		private var processFailedMessage: Option[Message] = None
+		override def hasNext: Boolean = openFolder.isDefined || queuedFailures.hasNext || openNextFolder()
+		
+		
+		// IMPLEMENTED  --------------------------
+		
+		@tailrec
+		override final def next(): Try[Message] = queuedFailures.nextOption() match {
+			// Case: A failure was queued => Returns the queued failure
+			case Some(error) => Failure(error)
+			case None =>
+				openFolder match {
+					// Case: A folder has been opened (as expected) => Reads the next message from the folder
+					case Some((folder, folderSize)) =>
+						val message = Try { folder.getMessage(nextMessageIndex) }
+						// Case: There remain messages to read => Advances the index
+						if (nextMessageIndex < folderSize)
+							nextMessageIndex += 1
+						// Case: All messages have now been read => Closes the folder
+						else
+							close(folder).failure.foreach(queue)
+						// Returns the message read result
+						message
+					// Case: A folder has not been opened (unexpected) => Attempts to open one
+					case None =>
+						// Case: Opening yielded something to return => Returns that item
+						if (openNextFolder())
+							next()
+						// Case: No more elements to return => Throws
+						else
+							throw new NoSuchElementException("next() called on an empty iterator")
+				}
+		}
+		
+		
+		// OTHER    -----------------------------
+		
+		/**
+		 * @return Closes the currently open folder, if applicable.
+		 *         Should be called before discarding this iterator, unless it has already been consumed
+		 */
+		def close(): Try[Unit] = openFolder match {
+			case Some((folder, _)) => close(folder)
+			case None => Success(())
+		}
+		
+		// Returns if this iterator may return new items afterwards
+		@tailrec
+		private def openNextFolder(): Boolean = source.nextOption() match {
+			// Case: Failed to read the next folder => Queues the encountered failure
+			case Some(Failure(error)) =>
+				queue(error)
+				true
+			// Case: Next folder is available
+			case Some(Success(folder)) =>
+				// Attempts to open the folder
+				Try { folder.open(if (deletionRule.canDelete) Folder.READ_WRITE else Folder.READ_ONLY) }
+					// Counts the messages in the folder
+					.flatMap { _ => Try { folder.getMessageCount } } match
+				{
+					// Case: Folder successfully opened
+					case Success(messageCount) =>
+						// Case: All messages in this folder are requested to be skipped
+						// => Closes the folder immediately
+						if (remainingSkipCount >= messageCount) {
+							remainingSkipCount -= messageCount
+							close(folder) match {
+								// Case: Closing succeeded => Moves to the next folder, if possible
+								case Success(_) => openNextFolder()
+								// Case: Closing failed => Prepares the closing error as the next element
+								case Failure(error) =>
+									queue(error)
+									true
+							}
+						}
+						// Case: There are messages to iterate => Prepares for the iteration
+						else {
+							nextMessageIndex = 1 + remainingSkipCount
+							remainingSkipCount = 0
+							openFolder = Some(folder -> messageCount)
+							true
+						}
+					// Case: Failed to open the folder => Prepares the error as the next element
+					case Failure(error) =>
+						queue(error)
+						true
+				}
+			// Case: No more folders available
+			case None => false
+		}
+		
+		private def close(folder: Folder) = {
+			openFolder = None
+			Try { folder.close(deletionRule.canDelete) }
+		}
+		
+		private def queue(error: Throwable) = queuedFailures :+= error
+	}
+	
+	private class MessageIterator(source: RawMessageIterator, deletionRule: DeletionRule) extends Iterator[Try[A]]
+	{
+		// ATTRIBUTES   ---------------------------
+		
+		// Finds the next non-skipped message & builder
+		private val nextMaterials = ResettableLazy {
+			source.findMapNext {
+				case Success(materials) => materials.map { Success(_) }
+				case Failure(error) => Some(Failure(error))
+			}
+		}
 		
 		
 		// IMPLEMENTED  ---------------------------
 		
-		override def hasNext = processFailedMessage.isEmpty && nextMaterials.value.isDefined
+		override def hasNext = nextMaterials.value.isDefined
 		
-		override def next() = nextMaterials.pop().get match {
+		override def next() = nextMaterials.pop()
+			.getOrElse { throw new NoSuchElementException("Called next() on an empty iterator") } match
+		{
 			// Case: Next message items could be successfully acquired => Processes message content
 			case Success((message, builder)) =>
 				// Reads the message content. If the primary content read fails,
 				// attempts to get around it by wrapping the message in another instance
 				// (Unable to load BODYSTRUCTURE exception workaround)
-				val content = Try { message.getContent } match
-				{
+				val content = Try { message.getContent } match {
 					// Case: Access succeeded => continues normally
 					case success: Success[AnyRef] => success
 					// Case: Initial access failed => Attempts the workaround if applicable
 					case failure: Failure[AnyRef] =>
-						message match
-						{
+						message match {
 							// Case: MimeMessage => Tries the workaround
 							case mimeMessage: MimeMessage =>
 								// If the workaround fails also, refers back to the original exception
@@ -256,11 +369,14 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 				val result = content
 					.flatMap { processContent(_, builder) }
 					.flatMap { _ => builder.result() }
+				// May delete the message afterwards
 				// Remembers the result
-				if (result.isSuccess)
-					processedMessagesBuilder += message
-				else
-					processFailedMessage = Some(message)
+				// Ignores deletion failures
+				if (result.isSuccess) {
+					if (deletionRule.shouldDeleteProcessed)
+						delete(message)
+				} else if (deletionRule.shouldDeleteFailed)
+					delete(message)
 				result
 			// Case: Message parsing failure => this fails also
 			case Failure(error) => Failure(error)
@@ -269,20 +385,11 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		
 		// OTHER    --------------------------------
 		
-		def popMessages() = {
-			val (skipped, rawFailed) = source.popMessages()
-			val failed = processFailedMessage.orElse(rawFailed)
-			processFailedMessage = None
-			(processedMessagesBuilder.result(), skipped, failed)
-		}
-		
-		private def processContent(content: AnyRef, builder: FromEmailBuilder[A]): Try[Unit] =
-		{
+		private def processContent(content: AnyRef, builder: FromEmailBuilder[A]): Try[Unit] = {
 			if (content == null)
 				Success(())
 			else
-				content match
-				{
+				content match {
 					case string: String => builder.append(string)
 					case stream: InputStream => builder.appendFrom(stream)
 					case multiPart: Multipart =>
@@ -290,8 +397,7 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 							.map { _.filterNot { _ == null } }
 							.flatMap { _.tryForeach { part =>
 								// Catches exceptions since almost every Java interface method throws
-								Try
-								{
+								Try {
 									// Handles attachments or other part content
 									val disposition = part.getDisposition
 									// Case: Attachment => Attaches it using the builder
@@ -312,71 +418,50 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		}
 	}
 	
-	private class RawMessageIterator(source: Iterator[Try[Message]],
-	                              makeBuilder: LazyEmailHeadersView => Option[FromEmailBuilder[A]])
+	// Returns Success(None) for messages that are skipped
+	private class RawMessageIterator(source: Iterator[Try[Message]], deletionRule: DeletionRule,
+	                                 makeBuilder: LazyEmailHeadersView => Option[FromEmailBuilder[A]])
 		extends Iterator[Try[Option[(Message, FromEmailBuilder[A])]]]
 	{
-		// ATTRIBUTES   -----------------------------
-		
-		private val skippedMessagesBuilder = new VectorBuilder[Message]()
-		private var failedMessage: Option[Message] = None
-		private var failed = false
-		
-		
 		// IMPLEMENTED  -----------------------------
 		
-		override def hasNext = !failed && source.hasNext
+		override def hasNext = source.hasNext
 		
-		override def next() =
-		{
+		override def next() = {
 			// Reads the next message (may fail)
-			source.next() match {
-				// Case: Message read => processes it
-				case Success(message) =>
-					Try {
-						// Reads the header information (lazily)
-						def subject = message.getSubject
-						def sentTime = message.getSentDate.toInstant
-						def sender = Option(message.getFrom) match
-						{
-							case Some(senders) => senders.mkString(", ")
-							case None => ""
-						}
-						def recipients =
-							Vector(RecipientType.TO, RecipientType.CC, RecipientType.BCC)
-								.flatMap { recipientType =>
-									Option(message.getRecipients(recipientType))
-										.filter { _.length > 0 }
-										.map { recipientType -> _.toVector.map { _.toString } }
-								}.toMap
-						def replyTo = Option(message.getReplyTo).flatMap { _.headOption }.map { _.toString }
-							.getOrElse("")
-						// Creates a builder, if necessary
-						makeBuilder(new LazyEmailHeadersView(sender, subject, sentTime, recipients, replyTo))
-					} match {
-						case Success(builder) =>
-							// Remembers skipped messages
-							if (builder.isEmpty)
-								skippedMessagesBuilder += message
-							Success(builder.map { message -> _ })
-						case Failure(error) =>
-							// Remembers last failed message
-							failedMessage = Some(message)
-							failed = true
-							Failure(error)
+			source.next().flatMap { message =>
+				Try {
+					// Reads the header information (lazily)
+					def subject = message.getSubject
+					def sentTime = message.getSentDate.toInstant
+					def sender = Option(message.getFrom) match {
+						case Some(senders) => senders.mkString(", ")
+						case None => ""
 					}
-				// Case: Message failed to be read => terminates
-				case Failure(error) =>
-					failed = true
-					Failure(error)
+					def recipients =
+						Vector(RecipientType.TO, RecipientType.CC, RecipientType.BCC)
+							.flatMap { recipientType =>
+								Option(message.getRecipients(recipientType))
+									.filter { _.length > 0 }
+									.map { recipientType -> _.toVector.map { _.toString } }
+							}.toMap
+					def replyTo = Option(message.getReplyTo).flatMap { _.headOption }.map { _.toString }
+						.getOrElse("")
+					// Creates a builder, if necessary
+					makeBuilder(new LazyEmailHeadersView(sender, subject, sentTime, recipients, replyTo))
+				} match {
+					case Success(builder) =>
+						// May delete skipped messages (ignores failures)
+						if (builder.isEmpty && deletionRule.shouldDeleteSkipped)
+							delete(message)
+						Success(builder.map { message -> _ })
+					case Failure(error) =>
+						// May delete failed messages (ignores failures)
+						if (deletionRule.shouldDeleteFailed)
+							delete(message)
+						Failure(error)
+				}
 			}
-		}
-		
-		def popMessages() =
-		{
-			val failed = failedMessage
-			failedMessage = None
-			skippedMessagesBuilder.result() -> failed
 		}
 	}
 }